+++ features/Context/FixturesContext.php
@@ -16,7 +16,6 @@ use Behat\Gherkin\Node\TableNode;
 use Doctrine\Common\Util\ClassUtils;
 use League\Flysystem\MountManager;
 use OAuth2\OAuth2;
-use Oro\Bundle\UserBundle\Entity\Role;
 use PHPUnit\Framework\Assert;
 use Pim\Behat\Context\FixturesContext as BaseFixturesContext;
 use Pim\Bundle\CatalogBundle\Entity\AttributeOption;
@@ -44,6 +43,7 @@ use Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductCsvIm
 use Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductModelCsvImport;
 use Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleCsvExport;
 use Pim\Component\ReferenceData\Model\ReferenceDataInterface;
+use Pim\Component\User\Model\Role;
 
 /**
  * A context for creating entities
@@ -1557,7 +1557,7 @@ class FixturesContext extends BaseFixturesContext
     /**
      * @param string $userRoleName
      *
-     * @return \Oro\Bundle\UserBundle\Entity\Role
+     * @return Role
      *
      * @Then /^there should be a "([^"]+)" user role$/
      */
@@ -1569,7 +1569,7 @@ class FixturesContext extends BaseFixturesContext
     /**
      * @param string $roleLabel
      *
-     * @return \Oro\Bundle\UserBundle\Entity\Role
+     * @return Role
      */
     public function getRole($roleLabel)
     {
