+++ src/Oro/Bundle/UserBundle/Entity/Group.php
@@ -8,6 +8,7 @@ use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation\Exclude;
 use JMS\Serializer\Annotation\Type;
 use Pim\Component\User\Model\GroupInterface;
+use Pim\Component\User\Model\Role;
 
 /**
  * @ORM\Entity(repositoryClass="Oro\Bundle\UserBundle\Entity\Repository\GroupRepository")
