+++ src/Oro/Bundle/UserBundle/Form/Type/AclRoleType.php
@@ -4,8 +4,8 @@ namespace Oro\Bundle\UserBundle\Form\Type;
 
 use Oro\Bundle\SecurityBundle\Form\Type\AclPrivilegeType;
 use Oro\Bundle\SecurityBundle\Form\Type\PrivilegeCollectionType;
-use Oro\Bundle\UserBundle\Entity\Role;
 use Pim\Bundle\EnrichBundle\Form\Type\EntityIdentifierType;
+use Pim\Component\User\Model\Role;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
