+++ src/Oro/Bundle/UserBundle/Tests/Unit/Entity/RoleTest.php
@@ -2,7 +2,7 @@
 
 namespace Oro\Bundle\UserBundle\Tests\Entity;
 
-use Oro\Bundle\UserBundle\Entity\Role;
+use Pim\Component\User\Model\Role;
 
 class RoleTest extends \PHPUnit_Framework_TestCase
 {
